apps/ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain>=0.2.0
langchain-google-genai>=0.0.6
google-generativeai==0.3.2
pymongo==4.6.0
sentence-transformers==2.2.2
python-dotenv>=1.0.0
pypdf2==3.0.1
pydantic>=2.9.0
pydantic-settings>=2.6.0
numpy>=2.0.0
redis==5.0.1

